+++ Source/Libraries/Common/Common.csproj
@@ -54,6 +54,7 @@
     <Compile Include="Forms\LogConsoleForm.Designer.cs">
       <DependentUpon>LogConsoleForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="ConsoleCopy.cs" />
     <Compile Include="Logging.cs" />
     <Compile Include="Extensions\RandomExtensions.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
